Dll creation support for Trans
This patch allows transcc to create dlls.
It introduces the new keyword "dllfunction" which triggers Dll creation
when used. dllfunctions are excluded from deadcode elimination and a new
flag disguises them from functions. dllfunctions have raw names and are
therefore not 'munged' this is necessary to satisfy plugin interfaces
for third-party applications. when a dll function is used the
correspondet file will be a *.dll instead of an exe.  The GC_MODE should
automatically change to Mode 2 when a dllfunction is used. There is one
remaining quirk: dllfunctions are still exported to the glfw target even
if it makes no sense and could break something. I don't know what would
be the best way to solve this?!