Mojo.Audio: Add GetChannelVolume(channel) and GetMusicVolume()
I do not know about that "If MOJO_VERSION_X ... Import mojox.audio", so
I am unsure what is done there (am only using the free monkey-version).

As there is no "MaxChannels"-value stored in the base audio.monkey, I
checked the available channels in GLFW and HTML5 and both had 33
channels (32 sfx and 1 music).
My implementation just hooks into the volume-setters and allows for
easier retrieval of current volumes.

Why? This is needed for situations in which multiple objects adjusts
volumes without knowing the other manipulators. It is similar to
GetAlpha() and SetAlpha().